+++ addon/popup.js
@@ -1,5 +1,5 @@
 /* global React ReactDOM */
-import {sfConn, apiVersion} from "./inspector.js";
+import {sfConn, apiVersion, sessionError} from "./inspector.js";
 import {getAllFieldSetupLinks} from "./setup-links.js";
 import {setupLinks} from "./links.js";
 
@@ -94,7 +94,7 @@ class App extends React.PureComponent {
     exportArg.set("host", sfHost);
     importArg.set("host", sfHost);
     limitsArg.set("host", sfHost);
-    if (e.contextSobject) {
+    if (e.contextSobject && localStorage.getItem("useSObjectContextOnDataImportLink") !== "false") {
       let query = "SELECT Id FROM " + e.contextSobject;
       if (e.contextRecordId && (e.contextRecordId.length == 15 || e.contextRecordId.length == 18)) {
         query += " WHERE Id = '" + e.contextRecordId + "'";
@@ -183,6 +183,21 @@ class App extends React.PureComponent {
       });
     }
   }
+  getBannerUrlAction(sessionError, sfHost, clientId, browser) {
+    let url;
+    let title;
+    let text;
+    if (sessionError !== "Session expired or invalid"){
+      text = "Access Token Expired";
+      title = "Generate New Token";
+      url = `https://${sfHost}/services/oauth2/authorize?response_type=token&client_id=` + clientId + "&redirect_uri=" + browser + "-extension://" + chrome.i18n.getMessage("@@extension_id") + "/data-export.html";
+    } else {
+      text = "Expired session";
+      title = "Return to login page";
+      url = `https://${sfHost}/`;
+    }
+    return {title, url, text};
+  }
   render() {
     let {
       sfHost,
@@ -199,7 +214,7 @@ class App extends React.PureComponent {
     const browser = navigator.userAgent.includes("Chrome") ? "chrome" : "moz";
     const DEFAULT_CLIENT_ID = "3MVG9HB6vm3GZZR9qrol39RJW_sZZjYV5CZXSWbkdi6dd74gTIUaEcanh7arx9BHhl35WhHW4AlNUY8HtG2hs"; //Consumer Key of  default connected app
     const clientId = localStorage.getItem(sfHost + "_clientId") ? localStorage.getItem(sfHost + "_clientId") : DEFAULT_CLIENT_ID;
-    const oauthAuthorizeUrl = `https://${sfHost}/services/oauth2/authorize?response_type=token&client_id=` + clientId + "&redirect_uri=" + browser + "-extension://" + chrome.i18n.getMessage("@@extension_id") + "/data-export.html";
+    const bannerUrlAction = this.getBannerUrlAction(sessionError, sfHost, clientId, browser);
     return (
       h("div", {},
         h("div", {className: "slds-page-header slds-theme_shade popup-header header-light"},
@@ -248,17 +263,17 @@ class App extends React.PureComponent {
             }
           }
         }),
-        h("div", {id: "expiredTokenLink", className: "hide"},
+        h("div", {id: "invalidTokenBanner", className: "hide"},
           h(AlertBanner, {type: "warning",
-            bannerText: "Access Token Expired",
+            bannerText: bannerUrlAction.text,
             iconName: "warning",
             iconTitle: "Warning",
-            assistiveTest: "Access Token Expired",
+            assistiveTest: bannerUrlAction.text,
             onClose: null,
             link: {
-              text: "Generate New Token",
+              text: bannerUrlAction.title,
               props: {
-                href: oauthAuthorizeUrl,
+                href: bannerUrlAction.url,
                 target: linkTarget
               }
             }
@@ -284,16 +299,6 @@ class App extends React.PureComponent {
             ),
             h("div", {className: "slds-m-bottom_xx-small"},
               h("a", {ref: "apiExploreBtn", href: "explore-api.html?" + hostArg, target: linkTarget, className: "page-button slds-button slds-button_neutral"}, h("span", {}, "E", h("u", {}, "x"), "plore API"))
-            ),
-            h("div", {className: "slds-m-bottom_xx-small"},
-              h("a",
-                {
-                  ref: "generateToken",
-                  href: oauthAuthorizeUrl,
-                  target: linkTarget,
-                  className: !clientId ? "button hide" : "page-button slds-button slds-button_neutral"
-                },
-                h("span", {}, h("u", {}, "G"), "enerate Access Token"))
             ),
             // Workaround for in Lightning the link to Setup always opens a new tab, and the link back cannot open a new tab.
             inLightning && isInSetup && h("div", {className: "slds-m-bottom_xx-small"},
@@ -347,7 +352,7 @@ class App extends React.PureComponent {
               )
             )
           ),
-          h("div", {className: "slds-col slds-size_1-of-12 slds-text-align_right slds-icon_container slds-m-right_small", title: "Options"},
+          h("div", {id: "optionsBtn", className: "slds-col slds-size_1-of-12 slds-text-align_right slds-icon_container slds-m-right_small", title: "Options"},
             h("a", {ref: "optionsBtn", href: "options.html?" + hostArg, target: linkTarget},
               h("svg", {className: "slds-button slds-icon_x-small slds-icon-text-default slds-m-top_xxx-small", viewBox: "0 0 52 52"},
                 h("use", {xlinkHref: "symbols.svg#settings", style: {fill: "#9c9c9c"}})
@@ -966,7 +971,7 @@ class AllDataBoxSObject extends React.PureComponent {
     let {selectedValue, recordIdDetails} = this.state;
     return (
       h("div", {},
-        h(AllDataSearch, {ref: "allDataSearch", onDataSelect: this.onDataSelect, sobjectsList, getMatches: this.getMatches, inputSearchDelay: 0, placeholderText: "Record id, id prefix or object name", resultRender: this.resultRender}),
+        h(AllDataSearch, {ref: "allDataSearch", sfHost, onDataSelect: this.onDataSelect, sobjectsList, getMatches: this.getMatches, inputSearchDelay: 0, placeholderText: "Record id, id prefix or object name", title: "Click to show recent items", resultRender: this.resultRender}),
         selectedValue
           ? h(AllDataSelection, {ref: "allDataSelection", sfHost, showDetailsSupported, selectedValue, linkTarget, recordIdDetails, contextRecordId, isFieldsPresent})
           : h("div", {className: "all-data-box-inner empty"}, "No record to display")
@@ -1190,7 +1195,7 @@ class AllDataBoxOrg extends React.PureComponent {
               h("tbody", {},
                 h("tr", {},
                   h("th", {}, h("a", {href: "https://" + sfHost + "/lightning/setup/CompanyProfileInfo/home", title: "Company Information", target: linkTarget}, "Org Id:")),
-                  h("td", {}, orgInfo.Id)
+                  h("td", {}, orgInfo.Id.substring(0, 15))
                 ),
                 h("tr", {},
                   h("th", {}, h("a", {href: "https://status.salesforce.com/instances/" + orgInfo.InstanceName, title: "Instance status", target: linkTarget}, "Instance:")),
@@ -1603,9 +1608,9 @@ class AllDataSelection extends React.PureComponent {
   render() {
     let {sfHost, showDetailsSupported, contextRecordId, selectedValue, linkTarget, recordIdDetails, isFieldsPresent} = this.props;
     // Show buttons for the available APIs.
-    let buttons = Array.from(selectedValue.sobject.availableApis);
+    let buttons = selectedValue.sobject.availableApis ? Array.from(selectedValue.sobject.availableApis) : [];
     buttons.sort();
-    if (buttons.length == 0) {
+    if (buttons.length == 0 && !selectedValue.isRecent) {
       // If none of the APIs are available, show a button for the regular API, which will partly fail, but still show some useful metadata from the tooling API.
       buttons.push("noApi");
     }
@@ -1641,7 +1646,7 @@ class AllDataSelection extends React.PureComponent {
                   h("span", {}, (selectedValue.recordId) ? " / " + selectedValue.recordId : ""),
                 )
               ),
-              selectedValue.sobject.name.indexOf("__") == -1
+              selectedValue.sobject.name.indexOf("__") == -1 && selectedValue.sobject.availableApis
                 ? h("tr", {},
                   h("th", {}, "Doc:"),
                   h("td", {},
@@ -1694,18 +1699,18 @@ class AllDataRecordDetails extends React.PureComponent {
       return (
         h("table", {className},
           h("tbody", {},
-            h("tr", {},
+            recordIdDetails.recordName ? h("tr", {},
               h("th", {}, "Name:"),
               h("td", {},
                 h("a", {href: this.getRecordLink(sfHost, selectedValue.recordId), target: linkTarget}, recordIdDetails.recordName)
               )
-            ),
-            h("tr", {},
+            ) : null,
+            recordIdDetails.recordTypeName ? h("tr", {},
               h("th", {}, "RecType:"),
               h("td", {},
                 h("a", {href: this.getRecordTypeLink(sfHost, selectedValue.sobject.name, recordIdDetails.recordTypeId), target: linkTarget}, recordIdDetails.recordTypeName)
               )
-            ),
+            ) : null,
             h("tr", {},
               h("th", {}, "Created:"),
               h("td", {}, recordIdDetails.created + " (" + recordIdDetails.createdBy + ")")
@@ -1761,6 +1766,7 @@ class AllDataSearch extends React.PureComponent {
     this.state = {
       queryString: "",
       matchingResults: [],
+      recentItems: [],
       queryDelayTimer: null
     };
     this.onAllDataInput = this.onAllDataInput.bind(this);
@@ -1814,8 +1820,8 @@ class AllDataSearch extends React.PureComponent {
     this.setState({queryDelayTimer});
   }
   render() {
-    let {queryString, matchingResults} = this.state;
-    let {placeholderText, resultRender} = this.props;
+    let {queryString, matchingResults, recentItems} = this.state;
+    let {placeholderText, resultRender, sfHost} = this.props;
     return (
       h("div", {className: "input-with-dropdown"},
         h("input", {
@@ -1831,7 +1837,10 @@ class AllDataSearch extends React.PureComponent {
         h(Autocomplete, {
           ref: "autoComplete",
           updateInput: this.updateAllDataInput,
-          matchingResults: resultRender(matchingResults, queryString)
+          matchingResults: resultRender(matchingResults, queryString),
+          recentItems: resultRender(recentItems, queryString),
+          queryString,
+          sfHost
         }),
         h("svg", {viewBox: "0 0 24 24", onClick: this.onAllDataArrowClick},
           h("path", {d: "M3.8 6.5h16.4c.4 0 .8.6.4 1l-8 9.8c-.3.3-.9.3-1.2 0l-8-9.8c-.4-.4-.1-1 .4-1z"})
@@ -1873,7 +1882,32 @@ class Autocomplete extends React.PureComponent {
     this.setState({showResults: true, selectedIndex: 0, scrollToSelectedIndex: this.state.scrollToSelectedIndex + 1});
   }
   handleFocus() {
-    this.setState({showResults: true, selectedIndex: 0, scrollToSelectedIndex: this.state.scrollToSelectedIndex + 1});
+    let {recentItems} = this.props;
+    sfConn.rest("/services/data/v" + apiVersion + "/query/?q=SELECT+Id,Name,Type+FROM+RecentlyViewed+LIMIT+50").then(res => {
+      res.records.forEach(recentItem => {
+        recentItems.push({key: recentItem.Id,
+          value: {recordId: recentItem.Id, isRecent: true, sobject: {keyPrefix: recentItem.Id.slice(0, 3), label: recentItem.Type, name: recentItem.Name}},
+          element: [
+            h("div", {className: "autocomplete-item-main", key: "main"},
+              recentItem.Name,
+            ),
+            h("div", {className: "autocomplete-item-sub", key: "sub"},
+              h(MarkSubstring, {
+                text: recentItem.Type,
+                start: -1,
+                length: 0
+              }),
+              " • ",
+              h(MarkSubstring, {
+                text: recentItem.Id,
+                start: -1,
+                length: 0
+              })
+            )
+          ]});
+      });
+      this.setState({recentItems, showResults: true, selectedIndex: 0, scrollToSelectedIndex: this.state.scrollToSelectedIndex + 1});
+    });
   }
   handleBlur() {
     this.setState({showResults: false});
@@ -1929,9 +1963,14 @@ class Autocomplete extends React.PureComponent {
   onResultsMouseUp() {
     this.setState({resultsMouseIsDown: false});
   }
-  onResultClick(value) {
-    this.props.updateInput(value);
-    this.setState({showResults: false, selectedIndex: 0});
+  onResultClick(e, value) {
+    let {sfHost} = this.props;
+    if (value.isRecent){
+      window.open("https://" + sfHost + "/" + value.recordId, getLinkTarget(e));
+    } else {
+      this.props.updateInput(value);
+      this.setState({showResults: false, selectedIndex: 0});
+    }
   }
   onResultMouseEnter(index) {
     this.setState({selectedIndex: index, scrollToSelectedIndex: this.state.scrollToSelectedIndex + 1});
@@ -1964,34 +2003,36 @@ class Autocomplete extends React.PureComponent {
     }
   }
   render() {
-    let {matchingResults} = this.props;
+    let {matchingResults, recentItems} = this.props;
     let {
       showResults,
       selectedIndex,
       scrollTopIndex,
       itemHeight,
-      resultsMouseIsDown,
+      resultsMouseIsDown
     } = this.state;
     // For better performance only render the visible autocomplete items + at least one invisible item above and below (if they exist)
     const RENDERED_ITEMS_COUNT = 11;
     let firstIndex = 0;
-    let lastIndex = matchingResults.length - 1;
+    let autocompleteResults = recentItems.length > 0 ? recentItems : matchingResults;
+    let lastIndex = autocompleteResults.length - 1;
     let firstRenderedIndex = Math.max(0, scrollTopIndex - 2);
     let lastRenderedIndex = Math.min(lastIndex, firstRenderedIndex + RENDERED_ITEMS_COUNT);
     let topSpace = (firstRenderedIndex - firstIndex) * itemHeight;
     let bottomSpace = (lastIndex - lastRenderedIndex) * itemHeight;
     let topSelected = (selectedIndex - firstIndex) * itemHeight;
+
     return (
-      h("div", {className: "autocomplete-container", style: {display: (showResults && matchingResults.length > 0) || resultsMouseIsDown ? "" : "none"}, onMouseDown: this.onResultsMouseDown, onMouseUp: this.onResultsMouseUp},
+      h("div", {className: "autocomplete-container", style: {display: (showResults && (autocompleteResults.length > 0)) || resultsMouseIsDown ? "" : "none"}, onMouseDown: this.onResultsMouseDown, onMouseUp: this.onResultsMouseUp},
         h("div", {className: "autocomplete", onScroll: this.onScroll, ref: "scrollBox"},
           h("div", {ref: "selectedItem", style: {position: "absolute", top: topSelected + "px", height: itemHeight + "px"}}),
           h("div", {style: {height: topSpace + "px"}}),
-          matchingResults.slice(firstRenderedIndex, lastRenderedIndex + 1)
+          autocompleteResults.slice(firstRenderedIndex, lastRenderedIndex + 1)
             .map(({key, value, element}, index) =>
               h("a", {
                 key,
                 className: "autocomplete-item " + (selectedIndex == index + firstRenderedIndex ? "selected" : ""),
-                onClick: () => this.onResultClick(value),
+                onClick: (e) => this.onResultClick(e, value),
                 onMouseEnter: () => this.onResultMouseEnter(index + firstRenderedIndex)
               }, element)
             ),